+++ lightly_studio/src/lightly_studio/cli.py
@@ -8,6 +8,8 @@ from pathlib import Path
 import click
 
 import lightly_studio
+from lightly_studio.analytics import tracking
+from lightly_studio.analytics.tracking import LaunchSource
 from lightly_studio.database import db_manager
 from lightly_studio.evaluation.image_dataset_evaluate import ObjectDetectionEvaluationConfig
 
@@ -69,6 +71,10 @@ def quickstart(port: int | None, force_download: bool, no_browser: bool) -> None
         config=evaluation_config,
     )
 
+    tracking.track(
+        event=tracking.APP_LAUNCHED,
+        properties={"launch_source": LaunchSource.QUICKSTART},
+    )
     lightly_studio.start_gui(port=port, open_browser=not no_browser)
 
 
@@ -99,4 +105,8 @@ def gui(
     if db_file is not None and db_url is not None:
         raise click.UsageError("Options '--db-file' and '--db-url' are mutually exclusive.")
     db_manager.connect(db_file=db_file, db_url=db_url, must_exist=True)
+    tracking.track(
+        event=tracking.APP_LAUNCHED,
+        properties={"launch_source": LaunchSource.GUI},
+    )
     lightly_studio.start_gui(host=host, port=port)